refactor(api): isolate playlist job routes
